projects/Hash/requirements.txt
colorama
tqdm
pycryptodome
crcmod
tabulate

